refactor(utils): move tokens_to_base_units/base_units_to_tokens to CLI crate
Per review feedback from igamigo: these functions are only used by the CLI,
so they belong in bin/miden-cli rather than the client library.

Changes:
- Move TokenParseError, tokens_to_base_units, base_units_to_tokens from
  crates/rust-client/src/utils.rs to bin/miden-cli/src/utils.rs
- Add miden-protocol and miden-standards as CLI Cargo.toml dependencies
- Fix TokenParseError::AmountTooLarge message (remove 'u64' mention)
- Fix per-part validation loop: map IntErrorKind::PosOverflow to AmountTooLarge
  so an oversized integer part also returns AmountTooLarge, not ParseU64
- Update account.rs to import base_units_to_tokens from crate::utils
